+++ CHANGES.md
@@ -2,6 +2,7 @@
 
 ## 0.22.1 (2025-mm-dd)
 
+- FIX: Correct the logic behind saving a stack in uint16 in `product.stack`  
 - FIX: Fix a regression disallowing to pass a custom nodata to `product.stack`  
 
 ## 0.22.0 (2025-03-25)